refactor: videoRoom 컴포넌트를 sessionPage 페이지 컴포넌트로 변경
